+++ examples/json/parser_streaming.rs
@@ -5,7 +5,7 @@ use winnow::prelude::*;
 use winnow::{
   branch::alt,
   bytes::{any, none_of, one_of, tag, take, take_while},
-  character::f64,
+  character::float,
   combinator::{cut_err, rest},
   error::{ContextError, ParseError},
   input::Streaming,
@@ -54,7 +54,7 @@ fn json_value<'i, E: ParseError<Input<'i>> + ContextError<Input<'i>, &'static st
     null.value(JsonValue::Null),
     boolean.map(JsonValue::Boolean),
     string.map(JsonValue::Str),
-    f64.map(JsonValue::Num),
+    float.map(JsonValue::Num),
     array.map(JsonValue::Array),
     object.map(JsonValue::Object),
   ))(input)
